Added done button
Added a done button to create question bank div. This button is actually a form in disguise. The purpose of this form is to get the data from all the questions fields and convert it to JSON before posting it to a php script that will create a compressed file that can be downloaded by the user.