Created setup div
Created the setup div in game.php. This is the initial div that shows up when the game form is loaded with the appropriate data. In this div, the user can set the amount of time given to each team/player and the names for as many teams as are playing.